+++ resolwe_bio/tests/processes/test_pca.py
@@ -40,11 +40,11 @@ class PcaProcessorTestCase(KBBioProcessTestCase):
             'species': 'Dictyostelium discoideum'
         }
         pca = self.run_process('pca', inputs)
-        self.assertEqual(len(pca.process_warning), 1)
-        self.assertEqual(
-            pca.process_warning[0],
-            'Gene selection and filtering resulted in no genes. Please select different samples or genes.'
-        )
+        saved_json, test_json = self.get_json('pca_plot2.json.gz', pca.output['pca'])
+        self.assertAlmostEqualGeneric(test_json['flot']['data'], saved_json['flot']['data'])
+        self.assertAlmostEqualGeneric(test_json['explained_variance_ratios'], saved_json['explained_variance_ratios'])
+        self.assertAlmostEqualGeneric(test_json['components'], saved_json['components'])
+        self.assertEqual(len(pca.process_warning), 0)
 
     @with_resolwe_host
     @tag_process('pca')